refactor(user, sample): user 패키지를 sample 패키지로 변경 후 다시 user 패키지 생성
