+++ src/main/java/com/flab/marketgola/sample/exception/SampleErrorResult.java
@@ -1,4 +1,4 @@
-package com.flab.marketgola.user.exception;
+package com.flab.marketgola.sample.exception;
 
 import lombok.Data;
